# swagger/swagger.yaml
consumes:
- application/json
definitions:
  AerospikeAmcConnectionTreeResponse:
    description: User Connection Entity Tree (default view)
    example:
      id: 70f01ba5-b14f-47d9-8d69-c5b4e960d88b
      modules:
      - hash: 06508300d1dc7e308ce8998bb4fe7a12505aa1f9
        name: map_reduce.lua
        type: lua
      nodes:
      - host: 127.0.0.1:3000
        id: BB9EB6C71290C00
        namespaces:
        - name: Test
          sets:
          - indexes:
            - binName: id
              name: idx_test_customers_id
              type: NUMERIC
            name: Customers
      - host: 127.0.0.1:3000
        id: BB9EB6C71290C00
        namespaces:
        - name: Test
          sets:
          - indexes:
            - binName: id
              name: idx_test_customers_id
              type: NUMERIC
            name: Customers
    properties:
      id:
        description: Connection Id
        example: 70f01ba5-b14f-47d9-8d69-c5b4e960d88b
        pattern: '[a-f0-9]{8}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{12}'
        type: string
      modules:
        description: modules
        example:
        - hash: 06508300d1dc7e308ce8998bb4fe7a12505aa1f9
          name: map_reduce.lua
          type: lua
        items:
          $ref: '#/definitions/AerospikeAmcEntityModuleResponse'
        type: array
      nodes:
        description: Nodes
        example:
        - host: 127.0.0.1:3000
          id: BB9EB6C71290C00
          namespaces:
          - name: Test
            sets:
            - indexes:
              - binName: id
                name: idx_test_customers_id
                type: NUMERIC
              name: Customers
        - host: 127.0.0.1:3000
          id: BB9EB6C71290C00
          namespaces:
          - name: Test
            sets:
            - indexes:
              - binName: id
                name: idx_test_customers_id
                type: NUMERIC
              name: Customers
        items:
          $ref: '#/definitions/AerospikeAmcEntityNodeResponse'
        type: array
    required:
    - id
    title: 'Mediatype identifier: application/vnd.aerospike.amc.connection.tree.response+json;
      view=default'
    type: object
...
  AerospikeAmcSystemResponse:
    description: AMC Server System information (default view)
    example:
      version: "5.0"
    properties:
      version:
        description: AMC Version
        example: "5.0"
        type: string
    title: 'Mediatype identifier: application/vnd.aerospike.amc.system.response+json;
      view=default'
    type: object
  AerospikeAmcUserQueryResponse:
    description: User (default view)
    example:
      active: true
      fullName: Ada Lovelace
      notes: A Genius
      roles:
      - ops
      - dev
      username: ada
    properties:
      active:
        description: User account is active
        example: true
        type: boolean
      fullName:
        description: User's fullname
        example: Ada Lovelace
        type: string
      notes:
        description: Additional Notes
        example: A Genius
        type: string
      roles:
        description: AMC User Roles
        example:
        - ops
        - dev
        items:
          example: Aut voluptas.
          type: string
        type: array
      username:
        description: User Id
        example: ada
        type: string
    required:
    - username
    title: 'Mediatype identifier: application/vnd.aerospike.amc.user.query.response+json;
      view=default'
    type: object
  AuthenticateAuthPayload:
    example:
      password: password
      user: username
    properties:
      password:
        description: AMC User's Password
        example: password
        type: string
      user:
        description: AMC User Name (this is not the same as Database user name)
        example: username
        type: string
    required:
    - user
    - password
    title: AuthenticateAuthPayload
    type: object
  ConnectConnectionPayload:
    example:
      password: "123456"
      username: admin
    properties:
      password:
        description: Database User's Password
        example: "123456"
        type: string
      username:
        description: Database Username
        example: admin
        type: string
    required:
    - username
    - password
    title: ConnectConnectionPayload
    type: object
...
  SaveConnectionPayload:
    example:
      id: 70f01ba5-b14f-47d9-8d69-c5b4e960d88b
      name: Payments Cluster
      seeds:
      - host: 127.0.0.1
        port: 3000
        tlsName: ""
    properties:
      id:
        description: Connection Id
        example: 70f01ba5-b14f-47d9-8d69-c5b4e960d88b
        pattern: '[a-f0-9]{8}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{12}'
        type: string
      name:
        description: Connection Name
        example: Payments Cluster
        type: string
      seeds:
        description: Seeds
        example:
        - host: 127.0.0.1
          port: 3000
          tlsName: ""
        items:
          $ref: '#/definitions/NodeSeed'
        type: array
    required:
    - name
    - seeds
    title: SaveConnectionPayload
    type: object
  SaveUserPayload:
    example:
      active: true
      fullName: Ada Lovelace
      notes: A Genius
      password: password
      roles:
      - ops
      - dev
      username: ada
    properties:
      active:
        description: User account is active
        example: true
        type: boolean
      fullName:
        description: User's fullname
        example: Ada Lovelace
        type: string
      notes:
        description: Additional Notes
        example: A Genius
        type: string
      password:
        description: Password
        example: password
        type: string
      roles:
        description: 'Valid roles are: admin, ops, dev'
        example:
        - ops
        - dev
        items:
          example: Tempora similique.
          type: string
        type: array
      username:
        description: AMC User Id
        example: ada
        type: string
    required:
    - username
    - roles
    title: SaveUserPayload
    type: object
  error:
    description: Error response media type (default view)
    example:
      code: invalid_value
      detail: Value of ID must be an integer
      id: 3F1FKVRR
      meta:
        timestamp: 1.458609e+09
      status: "400"
    properties:
      code:
        description: an application-specific error code, expressed as a string value.
        example: invalid_value
        type: string
      detail:
        description: a human-readable explanation specific to this occurrence of the
          problem.
        example: Value of ID must be an integer
        type: string
      id:
        description: a unique identifier for this particular occurrence of the problem.
        example: 3F1FKVRR
        type: string
      meta:
        additionalProperties: true
        description: a meta object containing non-standard meta-information about
          the error.
        example:
          timestamp: 1.458609e+09
        type: object
      status:
        description: the HTTP status code applicable to this problem, expressed as
          a string value.
        example: "400"
        type: string
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
info:
  description: Aerospike Management Console is used to manage and monitor Aerospike
    Database Clusters.
  title: Aerospike Management Console
  version: ""
paths:
  /{filepath}:
    get:
      operationId: public#/*filepath
      parameters:
      - description: Relative file path
        in: path
        name: filepath
        required: true
        type: string
      responses:
        "200":
          description: File downloaded
          schema:
            type: file
        "404":
          description: File not found
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      - https
      summary: Download static_v5/
  /api/v1/amc/system:
    get:
      description: Get AMC Server Version and Information
      operationId: amc#system
      produces:
      - application/vnd.aerospike.amc.system.response+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/AerospikeAmcSystemResponse'
        "500":
          description: Internal Server Error
      schemes:
      - http
      - https
      summary: system amc
      tags:
      - amc
  /api/v1/auth/authenticate:
    post:
      description: Authenticate a device
      operationId: auth#authenticate
      parameters:
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/AuthenticateAuthPayload'
      responses:
        "204":
          description: No Content
          headers:
            Authorization:
              description: Generated JWT
              type: string
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
      schemes:
      - http
      - https
      summary: authenticate auth
      tags:
      - auth
  /api/v1/connections:
    get:
      description: |-
        Get the list of user connections

        Required security scopes:
          * `api:general`
      operationId: connection#query
      produces:
      - text/plain
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      schemes:
      - http
      - https
      security:
      - jwt:
        - api:general
      summary: query connection
      tags:
      - connection
    post:
      description: |-
        Register or Update a new connection for user

        Required security scopes:
          * `api:general`
      operationId: connection#save
      parameters:
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/SaveConnectionPayload'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      schemes:
      - http
      - https
      security:
      - jwt:
        - api:general
      summary: save connection
      tags:
      - connection
  /api/v1/connections/{id}:
    delete:
      description: |-
        Delete a connection

        Required security scopes:
          * `api:general`
      operationId: connection#delete
      parameters:
      - description: Connection Id
        in: path
        name: id
        pattern: '[a-f0-9]{8}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{12}'
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      schemes:
      - http
      - https
      security:
      - jwt:
        - api:general
      summary: delete connection
      tags:
      - connection
    post:
      description: |-
        Connect to the cluster and return the entity tree

        Required security scopes:
          * `api:general`
      operationId: connection#connect
      parameters:
      - description: Connection Id
        in: path
        name: id
        pattern: '[a-f0-9]{8}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{12}'
        required: true
        type: string
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/ConnectConnectionPayload'
      produces:
      - application/vnd.aerospike.amc.connection.tree.response+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/AerospikeAmcConnectionTreeResponse'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
      schemes:
      - http
      - https
      security:
      - jwt:
        - api:general
      summary: connect connection
      tags:
      - connection
  /api/v1/users:
    get:
      description: |-
        Get the list of users

        Required security scopes:
          * `api:admin`
      operationId: user#query
      produces:
      - text/plain
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      schemes:
      - http
      - https
      security:
      - jwt:
        - api:admin
      summary: query user
      tags:
      - user
    post:
      description: |-
        Register or Update a new user

        Required security scopes:
          * `api:admin`
      operationId: user#save
      parameters:
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/SaveUserPayload'
      produces:
      - application/vnd.aerospike.amc.user.query.response+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/AerospikeAmcUserQueryResponse'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      schemes:
      - http
      - https
      security:
      - jwt:
        - api:admin
      summary: save user
      tags:
      - user
  /api/v1/users/{username}:
    delete:
      description: |-
        Delete a user

        Required security scopes:
          * `api:admin`
      operationId: user#delete
      parameters:
      - in: path
        name: username
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      schemes:
      - http
      - https
      security:
      - jwt:
        - api:admin
      summary: delete user
      tags:
      - user
    get:
      description: |-
        Get the list of users

        Required security scopes:
          * `api:admin`
      operationId: user#get
      parameters:
      - in: path
        name: username
        required: true
        type: string
      produces:
      - application/vnd.aerospike.amc.user.query.response+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/AerospikeAmcUserQueryResponse'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      schemes:
      - http
      - https
      security:
      - jwt:
        - api:admin
      summary: get user
      tags:
      - user
...
produces:
- application/json
...
schemes:
- http
- https
securityDefinitions:
  jwt:
    description: |2-


      **Security Scopes**:
        * `api:cluster`: API access to cluster resources
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"